[NFC] Move FPBuiltinFnSelection pass to Transforms
This moves the FPBuiltinFnSelection pass to the
llvm/lib/Transforms/Scalar directory. This is needed to enable future
changes that will run this pass as part of the main pipeline for device
compilation. With this pass in CodeGen, the slibs build would fail.
