+++ llvm/lib/Transforms/Scalar/CMakeLists.txt
@@ -14,6 +14,7 @@ add_llvm_component_library(LLVMScalarOpts
   EarlyCSE.cpp
   FlattenCFGPass.cpp
   Float2Int.cpp
+  FPBuiltinFnSelection.cpp
   GuardWidening.cpp
   GVN.cpp
   GVNHoist.cpp
